+++ src/main/groovy/com/techexpo/menu/drink/SweetTea.java
@@ -2,13 +2,11 @@ package com.techexpo.menu.drink;
 
 import com.techexpo.menu.MenuItem;
 
-import java.math.BigDecimal;
-
 public class SweetTea extends MenuItem {
 
     public SweetTea() {
         super.name = "Sweet Tea";
-        super.price = BigDecimal.valueOf(1.59);
+        super.price = 1.59;
         super.calories = 220;
         super.carbs = 57;
         super.fat = 0;
