+++ src/main/groovy/com/techexpo/menu/entree/ChickenSandwich.java
@@ -2,13 +2,11 @@ package com.techexpo.menu.entree;
 
 import com.techexpo.menu.MenuItem;
 
-import java.math.BigDecimal;
-
 public class ChickenSandwich extends MenuItem {
 
     public ChickenSandwich() {
         super.name = "Chick-fil-A Chicken Sandwich";
-        super.price = BigDecimal.valueOf(3.05);
+        super.price = 3.05;
         super.calories = 440;
         super.carbs = 40;
         super.fat = 19;
