+++ src/main/groovy/com/techexpo/menu/side/Fries.java
@@ -2,13 +2,11 @@ package com.techexpo.menu.side;
 
 import com.techexpo.menu.MenuItem;
 
-import java.math.BigDecimal;
-
 public class Fries extends MenuItem {
 
     public Fries() {
         super.name = "Waffle Fries";
-        super.price = BigDecimal.valueOf(1.65);
+        super.price = 1.65;
         super.calories = 520;
         super.carbs = 63;
         super.fat = 27;
