+++ src/main/groovy/com/techexpo/menu/side/FruitCup.java
@@ -2,13 +2,11 @@ package com.techexpo.menu.side;
 
 import com.techexpo.menu.MenuItem;
 
-import java.math.BigDecimal;
-
 public class FruitCup extends MenuItem {
 
     public FruitCup() {
         super.name = "Fruit Cup";
-        super.price = BigDecimal.valueOf(2.75);
+        super.price = 2.75;
         super.calories = 80;
         super.carbs = 21;
         super.fat = 0;
